Fix: "인벤토리 오타 수정"
오타 수정

+++ routes/inventory.router.js
@@ -1,11 +1,11 @@
-import express from "express";
-import { prisma } from "../utils/prisma/prismaClient.js";
-import authMiddleware from "../middlewares/auth.middleware.js";
+import express from 'express';
+import { prisma } from '../utils/prisma/prismaClient.js';
+import authMiddleware from '../middlewares/auth.middleware.js';
 
 const router = express.Router();
 
 // 나의 계정에 있는 CharacterDB(보유 선수 목록) 조회하기
-router.get("/Inventory/Check", authMiddleware, async (req, res, next) => {
+router.get('/Inventory/Check', authMiddleware, async (req, res, next) => {
   //authMiddleware에서 인증한 유저 아이디 가져오기
   const userId = req.user.userId;
 
@@ -17,7 +17,7 @@ router.get("/Inventory/Check", authMiddleware, async (req, res, next) => {
   if (!inventory) {
     return res
       .status(404)
-      .json({ error: "해당 유저가 가진 인벤토리를 찾을 수 없습니다." });
+      .json({ error: '해당 유저가 가진 인벤토리를 찾을 수 없습니다.' });
   }
 
   async function upgradeCharacter(inventory) {
@@ -35,7 +35,7 @@ router.get("/Inventory/Check", authMiddleware, async (req, res, next) => {
       inventoryId: inventory.inventoryId,
       ...character,
       upgrade: upgrade.upgrade,
-      equip: inventory.equip
+      equip: inventory.equip,
     };
     for (let key in character) {
       if (upgrade[key]) {
@@ -53,15 +53,14 @@ router.get("/Inventory/Check", authMiddleware, async (req, res, next) => {
     promiseAll.push(upgradeCharacter(inventory[i]));
   }
   await Promise.all(promiseAll);
-  
+
   characters.sort((a, b) => a.inventoryId - b.inventoryId);
 
   return res.status(200).json({ ...characters });
 });
 
-
 // 다른 사람의 계정에 있는 CharacterDB(보유 선수 목록) 조회하기
-router.get("/Inventory/Check/:userId", async (req, res, next) => {
+router.get('/Inventory/Check/:userId', async (req, res, next) => {
   //url에서 유저 아이디 받아오기
   const { userId } = req.params;
 
@@ -73,7 +72,7 @@ router.get("/Inventory/Check/:userId", async (req, res, next) => {
   if (!inventory) {
     return res
       .status(404)
-      .json({ error: "해당 유저가 가진 인벤토리를 찾을 수 없습니다." });
+      .json({ error: '해당 유저가 가진 인벤토리를 찾을 수 없습니다.' });
   }
 
   async function upgradeCharacter(inventory) {
@@ -91,7 +90,7 @@ router.get("/Inventory/Check/:userId", async (req, res, next) => {
       inventoryId: inventory.inventoryId,
       ...character,
       upgrade: upgrade.upgrade,
-      equip: inventory.equip
+      equip: inventory.equip,
     };
     for (let key in character) {
       if (upgrade[key]) {
@@ -115,29 +114,38 @@ router.get("/Inventory/Check/:userId", async (req, res, next) => {
 });
 
 // 선수 방출하기
-router.delete('/Inventory/Delete/:id', authMiddleware, async(req, res, next) => {  
-  const id = req.params.id;
-  const user = req.user;
-  
-  const isExistPlayer = await prisma.inventory.findUnique({
-    where: {
-      inventoryId: +id,
-    },
-  });
-  if(!isExistPlayer){
-    return res.status(404).json({ message: '해당 선수를 찾을 수 없습니다.'});
-  }
-  if(isExistPlayer.userId !== user.userId){
-    return res.status(400).json({message: '올바른 계정으로 접속해주세요.'});
-  }
-  
-  await prisma.inventory.delete({
-    where: {
-      inventoryId: +id,
-    },
-  });
+router.delete(
+  '/Inventory/Delete/:id',
+  authMiddleware,
+  async (req, res, next) => {
+    const id = req.params.id;
+    const user = req.user;
 
-  return res.status(200).json( {messeage: `${id}번 선수 방출했습니다.`} );
-});
+    const isExistPlayer = await prisma.inventory.findUnique({
+      where: {
+        inventoryId: +id,
+      },
+    });
+    if (!isExistPlayer) {
+      return res.status(404).json({ message: '해당 선수를 찾을 수 없습니다.' });
+    }
+    if (isExistPlayer.equip !== null) {
+      return res.status(400).json({
+        message: '스쿼드에 등록 되어있는 선수입니다. 방출하실 수 없습니다.',
+      });
+    }
+    if (isExistPlayer.userId !== user.userId) {
+      return res.status(400).json({ message: '올바른 계정으로 접속해주세요.' });
+    }
+
+    await prisma.inventory.delete({
+      where: {
+        inventoryId: +id,
+      },
+    });
+
+    return res.status(200).json({ message: `${id}번 선수 방출했습니다.` });
+  }
+);
 
 export default router;
